pyproject.toml
[project]
name = "strands-multi-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "aws-opentelemetry-distro==0.10.0",
    "bedrock-agentcore>=0.1.2",
    "bedrock-agentcore-starter-toolkit>=0.1.6",
    "boto3>=1.40.12",
    "strands-agents>=1.5.0",
    "strands-agents-builder>=0.1.8",
    "strands-agents-tools>=0.2.4",
    "uv>=0.8.12",
]
